requirements.txt
uvicorn
fastapi
sqlalchemy
fastapi-jwt-auth
python-dotenv
python-socketio
websockets